Rename ClassMethod#original_method_defined_on_stubbee?
The old name was slightly confusing because we're actually interested in
whether the original method is defined on the stubbee or the stubbee's
metaclass in the case of `AnyInstanceMethod`. The new name better describes why
we were interested in whether the original method was defined on the stubbee
(or stubbee's metaclass); notably that it means our stub method has/will
replace it.